refactor : 사용되지 않는 코드 삭제
