+++ frontend/types/community.ts
@@ -1,0 +1,15 @@
+export type Community = {
+  id: number;
+  email: string;
+  userStar: number;
+  avatar: string;
+  category: string;
+  title: string;
+  content: string;
+  createdAt: string;
+  heart: number;
+  view: number;
+  tags: string[];
+  status: null;
+  comment: never[];
+};
