+++ frontend/types/types.ts
@@ -3,47 +3,5 @@
 export type DefaultObj = {
   [key: string]: string;
 };
-export type job = {
-  [key: string]: { want: number; current: number };
-};
 
-export type Project = {
-  id: number;
-  author: string;
-  start: Date;
-  end: Date;
-  tags: string[];
-  stacks: string[];
-  jobs: job[];
-  title: string;
-  content: string;
-  state: 1 | 2 | 3 | 4;
-  createAt: string;
-  view: number;
-  heart: number;
-  comment: never[];
-  position?: string;
-};
-
-export type PostState = {
-  id: number;
-  heart: boolean;
-  want: string;
-};
-
-export type Article = {
-  id: number;
-  email: string;
-  userStar: number;
-  avatar: string;
-  category: string;
-  title: string;
-  content: string;
-  createdAt: string;
-  heart: number;
-  view: number;
-  tags: string[];
-  status: null;
-  comment: never[];
-};
 export type ArrObj = { [key: string]: string[] };
